fix(*): rename env credentials provider
