+++ packages/core/src/credentials-provider/EnvCredentialProvider.ts
