+++ packages/core/src/credentials-provider/index.ts
@@ -1,2 +1,2 @@
 export * from './CredentialsProvider';
-export * from './DefaultCredentialsProvider';
+export * from './EnvCredentialProvider';
